代码合并

+++ report-biz/src/main/java/com/jaagro/report/biz/service/impl/SettleManageServiceImpl.java
@@ -236,7 +236,7 @@ public class SettleManageServiceImpl implements SettleManageService {
                             .setTotalWeight(returnAccumulativeGoodsDto.getTotalWeight())
                             .setTotalQuantity(returnAccumulativeGoodsDto.getTotalQuantity());
                 }
-            }else {
+            } else {
                 driverSettleFeeMonthly.setTotalAnomalyFee(BigDecimal.ZERO)
                         .setTotalFreight(BigDecimal.ZERO)
                         .setTotalQuantity(0)
@@ -245,8 +245,10 @@ public class SettleManageServiceImpl implements SettleManageService {
             }
             driverSettleFeeMonthlyList.add(driverSettleFeeMonthly);
         }
-        driverSettleFeeMonthlyMapper.deleteByReportTime(month);
-        driverSettleFeeMonthlyMapper.batchSettleFeeMonthInsert(driverSettleFeeMonthlyList);
+        if (!CollectionUtils.isEmpty(driverSettleFeeMonthlyList)) {
+            driverSettleFeeMonthlyMapper.deleteByReportTime(month);
+            driverSettleFeeMonthlyMapper.batchSettleFeeMonthInsert(driverSettleFeeMonthlyList);
+        }
     }
 
     @Override
@@ -385,8 +387,10 @@ public class SettleManageServiceImpl implements SettleManageService {
             }
             customerSettleFeeMonthlyList.add(settleFeeMonthly);
         }
-        customerSettleFeeMonthlyMapper.delByReportTime(month);
-        customerSettleFeeMonthlyMapper.batchInsert(customerSettleFeeMonthlyList);
+        if (!CollectionUtils.isEmpty(customerSettleFeeMonthlyList)) {
+            customerSettleFeeMonthlyMapper.delByReportTime(month);
+            customerSettleFeeMonthlyMapper.batchInsert(customerSettleFeeMonthlyList);
+        }
     }
 
     /**
